pythonProject2/requirements.txt
# 机械零件同心度检测系统 - 依赖包清单
# Python 3.8+ 版本

# 核心计算机视觉库
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# GUI界面开发
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2

# 科学计算与数据处理
numpy==1.24.3
pandas==2.0.3
scipy==1.11.2

# 数据可视化
matplotlib==3.7.2

# 配置文件处理
pyyaml==6.0

# 图像处理
Pillow==10.0.0

# 报告生成
openpyxl==3.1.2
reportlab==4.0.4

# 日志与系统监控
psutil==5.9.6

# 测试框架
pytest==7.4.0
pytest-qt==4.2.0

# 开发工具
black==23.9.1
flake8==6.1.0
autopep8==2.0.4

# 其他工具
tqdm==4.66.1
scikit-image==0.21.0

# 可选：深度学习支持（如需）
# torch==2.0.1
# torchvision==0.15.2

# 可选：数据库支持（如需）
# sqlalchemy==2.0.19
# pymysql==1.1.0

# 可选：网络通信（如需）
# requests==2.31.0
# websockets==12.0

# 基础依赖
PyQt5>=5.15.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0

# 报告生成相关
reportlab>=3.6.0
python-docx>=0.8.11
pyyaml>=6.0

# 其他工具
psutil>=5.8.0  # 用于性能监控